LayoutTests/platform/qt/fast/css/font-face-descending-unicode-range-expected.txt

Interval unicode-ranges that descend are invalid and omitted. Without any valid ranges, the descriptor is omitted. As a result, the @font-face rule in this file is applied to all characters. This test passes if the following i's and X's are shown in Courier and they have the same length on screen. As of r62096, opening this file causes a crash instead.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

iiiiiiiiii
XXXXXXXXXX
FAIL document.getElementById('i_s').offsetWidth should be 112. Was 42.
PASS successfullyParsed is true

TEST COMPLETE
